+++ app/routes/topic.py
@@ -1,6 +1,9 @@
 from flask import Blueprint, jsonify, request, abort
+
+from app.middleware.auth import jwt_required
+from app.middleware.permissions import check_permission, grant_permission
 from app.models.topic import Topic
-from sqlalchemy import func
+from sqlalchemy import func, select
 from utils.text import limit_whitespace
 from app.extensions import db
 
@@ -8,8 +11,12 @@ topic_bp = Blueprint("topic", __name__, url_prefix="/topic")
 
 
 @topic_bp.route("", methods=["GET"])
-def all_topic():
-    result = Topic.query.order_by(func.coalesce(Topic.last_update, Topic.created_at)).all()
+@jwt_required()
+def all_topic(current_user):
+    selector = select(Topic).where(Topic.users.contains(current_user))
+    order = func.coalesce(Topic.last_update, Topic.created_at)
+    result = db.session.scalars(selector.order_by(order)).all()
+
     topics = {
         "no_of_topic": len(result),
         "topics": [topic.to_dict() for topic in result]
@@ -18,7 +25,8 @@ def all_topic():
 
 
 @topic_bp.route("", methods=["POST"])
-def add_topic():
+@jwt_required()
+def add_topic(current_user):
     data = request.form.to_dict()
     title = limit_whitespace(data.get("title"))
     emoji = limit_whitespace(data.get("emoji"))
@@ -27,33 +35,47 @@ def add_topic():
         return abort(400, "You must provide 'title' and 'emoji' for your new topic")
 
     new_topic = Topic(
-        title=limit_whitespace(title),
+        title=title,
         emoji=emoji,
+        creator=current_user
     )
     new_topic.progression_calc()
     db.session.add(new_topic)
+    db.session.flush()
+    grant_permission(current_user, new_topic, "creator")
     db.session.commit()
+    topic = db.session.scalar(select(Topic).where(Topic.id == new_topic.id))
+
+    if topic is None:
+        return abort(500, "seems like the topic was not created, please consult developer")
     if include:
-        topic = Topic.query.filter_by(id=new_topic.id).first()
         return jsonify({"success": f"successfully add new topic",
                         "topic": topic.to_dict()}), 201
     return jsonify({"success": f"successfully add new topic", "topic_id": new_topic.id}), 201
 
 
 @topic_bp.route("", methods=["PUT"])
-def update_topic():
+@jwt_required()
+def update_topic(current_user):
     data = request.form.to_dict()
     topic_id = limit_whitespace(data.get("topic_id"))
+
     try:
         topic_id = int(topic_id)
     except ValueError:
         return abort(400, "invalid 'topic_id' provided")
+
+    topic = db.session.scalar(select(Topic).where(Topic.id == topic_id))
+
     new_title = limit_whitespace(data.get("title"))
     new_emoji = limit_whitespace(data.get("emoji"))
-    topic = Topic.query.filter_by(id=topic_id).first()
     include = limit_whitespace(data.get("include", "false")).lower() == "true"
+
     if topic is None:
         return abort(404, f"no topic with id={topic_id} to update")
+
+    check_permission(current_user, topic, ["creator", "moderator"])
+
     if [new_title, new_emoji] == [None, None]:
         return abort(400, "you must provide 'title' or 'emoji' to update")
 
@@ -73,11 +95,12 @@ def update_topic():
 
 
 @topic_bp.route("", methods=["DELETE"])
-def delete_topic():
+@jwt_required()
+def delete_topic(current_user):
     data = request.form.to_dict()
     topic_id = limit_whitespace(data.get("topic_id"))
     topic_title = limit_whitespace(data.get("title"))
-    remove_tasks = limit_whitespace(data.get("delete_tasks", "true")) == "true"
+
     try:
         topic_id = int(topic_id)
     except ValueError:
@@ -85,30 +108,19 @@ def delete_topic():
     if topic_id is None:
         return abort(400, f"you must provide {topic_id} to remove")
 
-    topic = Topic.query.filter_by(id=topic_id).first()
+    topic = db.session.scalar(select(Topic).where(Topic.id == topic_id))
+    check_permission(current_user, topic, ["creator"])
+
     if topic is None:
         return abort(404, f"no topic with id={topic_id} found")
     if topic_title != topic.title:
         return abort(400, "topic title is not match")
 
-    removed_tasks = 0
-    orphaned_tasks = None
-    if remove_tasks:
-        for task in topic.tasks:
-            db.session.delete(task)
-            removed_tasks += 1
-    else:
-        orphaned_tasks = [task.to_dict() for task in topic.tasks]
     db.session.delete(topic)
     db.session.commit()
 
     response = {
         "success": "successfully removed a topic and all tasks associated with it",
-        "removed_topic": {"id": topic.id, "title": topic.title},
-        "tasks_removed": removed_tasks
+        "removed_topic": {"id": topic.id, "title": topic.title}
     }
-
-    if orphaned_tasks is not None:
-        response["success"] = "successfully removed a topic and there are orphaned tasks left"
-        response["orphaned_tasks"] = orphaned_tasks
-    return jsonify(response)
+    return jsonify(response), 200
